pyproject.toml
[project]
name = "on-chain-data-testing-project"
version = "0.1.0"
description = ""
authors = [
    {name = "LJ Li",email = "liliangjya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "apify-client (>=1.9.2,<2.0.0)",
    "solana (>=0.36.6,<0.37.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "base58 (>=2.1.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
